requirements.txt
PySide6>=6.5
Pillow>=10.0.0
watchdog>=3.0.0
imageio>=2.31.0
imageio-ffmpeg>=0.4.9
numpy>=1.24.0
optional
